Move I/O out of implementation file validation
